requirements.txt
absl_py==2.1.0
accelerate==0.31.0
gcsfs==2024.10.0
ipython==8.12.3
jax==0.4.37
jsonlines==4.0.0
loguru==0.7.2
ml_collections==0.1.1
openai==1.57.2
packaging==24.2
pandas==2.2.3
ray==2.35.0
regex==2024.9.11
Requests==2.32.3
setuptools==75.1.0
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
transformers==4.43.3
tyro==0.9.2
vllm==0.5.5
vllm-flash-attn==2.6.1
wandb==0.19.0
tensorboard==2.18.0
flax==0.10.2
